+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # X Changelog
 
+## 1.0.2
+
+### Testing
+
+- Use `jest-dom` for testing DOM to improve tests
+
+### x
+
+- Add `style()` function to set/get styles from a DOM node
+- Add `hover()` event listener
+- Improve `click()` for touch devices
+- Add `on()` for custom event listeners
+- Add `rmOn()` to remove event listeners created with `on()`
+
 ## 1.0.1
 
 ### Testing